fix issue with reference

+++ gbg_program/_Save_Recall.ino
@@ -543,9 +543,9 @@ void saveSettings()
   EEPROMwrite(addressW, SPEED_KNOB_SLOW_VAL);
   EEPROMwrite(addressW, SPEED_KNOB_FAST_VAL);
   EEPROMwrite(addressW, SCALE_ACCEL_WITH_SPEED);
-  EEPROMwrite(addressW, (byte)SPEED_KNOB_PIN);
-  EEPROMwrite(addressW, (byte)JOY_X_PIN);
-  EEPROMwrite(addressW, (byte)JOY_Y_PIN);
+  EEPROMwrite(addressW, SPEED_KNOB_PIN);
+  EEPROMwrite(addressW, JOY_X_PIN);
+  EEPROMwrite(addressW, JOY_Y_PIN);
 #ifdef IS_PCB
   EEPROMwrite(addressW, SWAP_MOTORS);
 #else
@@ -638,9 +638,9 @@ void recallSettings()
   EEPROMread(addressR, SPEED_KNOB_SLOW_VAL);
   EEPROMread(addressR, SPEED_KNOB_FAST_VAL);
   EEPROMread(addressR, SCALE_ACCEL_WITH_SPEED);
-  EEPROMread(addressR, (byte)SPEED_KNOB_PIN);
-  EEPROMread(addressR, (byte)JOY_X_PIN);
-  EEPROMread(addressR, (byte)JOY_Y_PIN);
+  EEPROMread(addressR, SPEED_KNOB_PIN);
+  EEPROMread(addressR, JOY_X_PIN);
+  EEPROMread(addressR, JOY_Y_PIN);
 #ifdef IS_PCB
   EEPROMread(addressR, SWAP_MOTORS);
 #else
